reorder main and app so that main has all wrappers and functionalities for the app
