+++ studentcompass.client/src/App.tsx
@@ -1,21 +1,16 @@
 import "./App.css";
-import { theme } from "./utils/themes";
+import React from "react";
+import { Box } from "@mui/material";
 import Navbar from "./components/Navbar";
-import { ToastContainer } from "react-toastify";
-import { Box, ThemeProvider } from "@mui/material";
-import { BrowserRouter } from "react-router-dom";
 
 function App() {
   return (
-    <BrowserRouter>
-      <ThemeProvider theme={theme}>
-        <Navbar />
-        <Box className={boxClassName} sx={boxStyle}>
-          <p>Content</p>
-        </Box>
-        <ToastContainer style={toastStyle} />
-      </ThemeProvider>
-    </BrowserRouter>
+    <React.Fragment>
+      <Navbar />
+      <Box className={boxClassName} sx={boxStyle}>
+        <p>Content</p>
+      </Box>
+    </React.Fragment>
   );
 }
 
@@ -31,10 +26,4 @@ const boxStyle = {
   overflow: "auto",
   flexDirection: "column",
   minHeight: "calc(100vh - 5rem)",
-};
-
-const toastStyle = {
-  maxWidth: 700,
-  marginTop: 67,
-  width: "auto",
 };
